pyproject.toml
[project]
name = "face-verification-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "absl-py==2.3.1",
    "attrs==25.3.0",
    "cffi==1.17.1",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "et-xmlfile==2.0.0",
    "flatbuffers==25.2.10",
    "fonttools==4.58.5",
    "jax==0.6.2",
    "jaxlib==0.6.2",
    "kiwisolver==1.4.8",
    "matplotlib==3.10.3",
    "mediapipe==0.10.14",
    "ml-dtypes==0.5.1",
    "numpy==2.2.6",
    "opencv-contrib-python==4.12.0.88",
    "opencv-python==4.12.0.88",
    "openpyxl==3.1.5",
    "opt-einsum==3.4.0",
    "packaging==25.0",
    "pandas==2.3.1",
    "pillow==11.3.0",
    "pip==25.1.1",
    "protobuf==4.25.8",
    "pycparser==2.22",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "scipy==1.16.0",
    "setuptools==65.5.0",
    "six==1.17.0",
    "sounddevice==0.5.2",
    "tzdata==2025.2",
]
